Consolodating types
